+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     library-level analysis_options.yaml instead.
   * Fixed "dart analyze" warnings about unused imports.
   * Fixed validation of function overloads for C++, Java, and Swift generators; and related issues.
+  * Fixed compilation issues in Java and Swift when two lambdas have the same signature.
 
 ## 9.3.5
 Release date: 2021-07-28